web/requirements.txt
Django==4.1.4
django-redis==5.2.0
djangorestframework==3.14.0
django-filter==22.1
pyjwt==2.5.0

gunicorn==20.1.0
whitenoise==6.2.0

psycopg2-binary==2.9.5

redis==4.4.0
celery==5.2.7

flake8==6.0.0
isort==5.10.1
black==22.10.0

djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
pika==1.3.1